pyproject.toml
[project]
name = "book-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikhail Osintsev",email = "a89150242288@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "pypdf (>=6.2.0,<7.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "ebooklib (>=0.20,<0.21)",
    "numpy (>=2.3.4,<3.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "langchain (>=1.0.5)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "pyopenssl (==23.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.7"
jupyter = "^1.1.1"

